% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distTraitCondR.R
\name{calcDistLROPostBreedingNoEnv}
\alias{calcDistLROPostBreedingNoEnv}
\title{Distribution of lifetime reproductive output}
\usage{
calcDistLROPostBreedingNoEnv(
  Pmat,
  Fmat,
  c0,
  maxClutchSize,
  maxLRO,
  Fdist = "Poisson"
)
}
\arguments{
\item{Pmat}{The survival/growth transition matrix.
Pmat[i,j] is the probability of transitioning from
state j to state i.}

\item{Fmat}{The fecundity matrix.  Fmat[i,j]
is the expected number of state i offspring from a state j
parent.}

\item{c0}{A vector specifying the offspring state distribution: c0[j] is
the probability that an individual is born in state j}

\item{maxClutchSize}{The maximum clutch size to consider}

\item{maxLRO}{The maximum LRO to consider}

\item{Fdist}{The clutch size distribution.  The recognized options are
"Poisson" and "Bernoulli".  Optional.  The default value is "Poisson".}
}
\value{
The distribution of lifetime reproductive output
}
\description{
Calculates the distribution of lifetime reproductive output (LRO) the
absence of environmental variation.  Assumes a post-breeding census
(reproductiion happens after survival and growth).
}
\details{
This function calculates the distribution of LRO by
cross-classifying states by stage and number of offspring produced
so far and calculating the state distribution at death.  If s(z')
is the survival probability for cross-classified state z', M(z', z)
is the probability of transitioning from cross-classified state z
to cross-classified state z', and N is the fundamental matrix for
M: N = (I - M)^\{-1\}, then the state distribution at death,
conditional on starting in state z, is (1 - s(z')) * N(z', z),
where * denotes a Hadamard product (element-by-element
multiplication, not matrix multiplication).  (See, e.g., eq. 3.2.8
in Data-driven Modeling of Structured Populations: A Practical
Guide to the Integral Projection Model, by Stephen P. Ellner, Dylan
Z. Childs and Mark Rees, ed. 1, 2015.)  This method implicitly
assumes that reproduction happens after survival and growth, i.e. a
post-breeding census.
}
\examples{
Pmat = matrix (c(0, 0.3, 0, 0, 0, 0.5, 0, 0, 0.5), 3, 3)
Fmat = matrix (0, 3, 3); Fmat[1,] = 0:2
Q = matrix (1/2, 2, 2)
c0 = c(1,0,0)
maxClutchSize = 10
maxLRO = 30
out = calcDistLROPostBreedingNoEnv (Pmat, Fmat, c0, maxClutchSize,
  maxLRO)
}
