% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_utilities.R
\name{calcDistOffspringCohort}
\alias{calcDistOffspringCohort}
\title{The distribution of offspring types in a cohort produced at the stable
population distribution}
\usage{
calcDistOffspringCohort(Amat, Fmat)
}
\arguments{
\item{Amat}{An \eqn{n \times n} matrix containing all of the transition and
reproductive rates that project the population from time \eqn{t} to
\eqn{t+1}. Generally given by \code{Umat+Fmat}.}

\item{Fmat}{An \eqn{n \times n} matrix of per capita rates of reproduction
rates by any of the \eqn{n} states into offspring (new individuals of any
of the \eqn{n} states)}
}
\value{
A vector of length \eqn{n} that contains the proportion of
individuals in each of the \eqn{n} states in a cohort of offspring
}
\description{
If adults in a population can produce multiple types of offspring, then life
can start in multiple states. Therefore, calculations of population-level
mean, variance, and skewness in lifespan or lifetime reproductive output must
take into account this variation in starting states. We often call this the
"mixing distribution" and a reasonable standard mixing distribution is the
distribution of offspring types in a cohort produced when the population is
at its stable population distribution.
}
\examples{
Umat<- matrix(c(0.5, 0.1, 0.1, 0, 0.5, 0.3, 0, 0, 0.8), ncol=3)
Fmat<- matrix(c(0.2, 0, 0, 1, 0.5, 0, 5, 3, 0), ncol=3)
Amat<- Umat+Fmat
mixdist<- calcDistOffspringCohort(Amat, Fmat)
}
