% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioningUtilities.R
\name{calcMoments}
\alias{calcMoments}
\title{Calculates the mean, the 2nd and 3rd non-central moments, and the skewness of
a lifetime reward}
\usage{
calcMoments(M, R1, R2, R3)
}
\arguments{
\item{M}{The transition matrix}

\item{R1}{A matrix.  R1[i,j] is the expected reward (e.g. reproductive
output) that occurs when making a transition from state j to state i.
Reward matrices include the "dead" state: if there are n states, R1 should
be size n+1 x n+1.}

\item{R2}{A matrix.  R2[i,j] is the expected squared reward (e.g.
reproductive output) that occurs when making a transition from state j to
state i.  Reward matrices include the "dead" state: if there are n states,
R2 should be size n+1 x n+1.}

\item{R3}{A matrix.  R3[i,j] is the expected cubed reward (e.g.
reproductive output) that occurs when making a transition from state j to
state i.  Reward matrices include the "dead" state: if there are n states,
R3 should be size n+1 x n+1.}
}
\value{
A list containing
\itemize{
\item rho1Vec: A vector whose jth entry gives the expected lifetime
reward for a trajectory starting in state j
\item mu2Vec: A vector whose jth entry gives the central 2nd moment
(variance) of lifetime reward for a trajectory starting in state j
\item mu3Vec: A vector whose jth entry gives the central 3rd moment of
lifetime reward for a trajectory starting in state j
\item skewnessVec: A vector whose jth entry gives the skewness (3rd
central moment / variance^3/2) of lifetime reward for a trajectory starting
in state j
}
}
\description{
Given a transition matrix and appropriate reward matrices, calculates the
mean, the 2nd and 3rd non-central moments, and the skewness of some quantity
(a "reward") over the stochastic trajectories generated by the transition
matrix.  If the reward matrices are for reproductive output, then this code
calculates moments of lifetime reproductive output.
}
\examples{
## Transition matrix: M\[i,j\] is the probability of transitioning
##  from state j to state i
M = matrix (0.1, 3, 3)
## Fecundity matrix: F\[i,j\] is the expected number of size i offspring
## producted in one time step for a parent of size j
F = matrix (0, 3, 3)
F[1,] = c(0, 1, 2)
## Clutch size is Poisson distributed
R1 = matrix (0, 4, 4)
for (j in 1:3)
  R1[,j] = sum(F[,j])
R2 = R1 + R1^2
R3 = R1 + R1^2 + R1^3
out = calcMoments (M, R1, R2, R3)
}
