% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distTraitCondR.R
\name{distLifespanCondR2}
\alias{distLifespanCondR2}
\title{Distribution of lifespan conditional on LRO}
\usage{
distLifespanCondR2(
  Plist,
  Flist,
  Q,
  c0,
  maxClutchSize,
  maxAge,
  percentileCutoff = 0.99,
  Fdist = "Poisson"
)
}
\arguments{
\item{Plist}{A list of survival/growth transition matrices.
Plist[[q]][i,j] is the probability of transitioning from
state j to state i in environment q.}

\item{Flist}{A list of fecundity matrices.  Flist[[q]][i,j]
is the expected number of state i offspring from a state j parent
in environment q}

\item{Q}{The environment transition matrix.  Q[i,j] is the
probability of transitioning from environment j to environment i.}

\item{c0}{A vector specifying the offspring state distribution:
c0[j] is the probability that an individual is born in state j}

\item{maxClutchSize}{The maximum clutch size to consider}

\item{maxAge}{The maximum attainable age}

\item{percentileCutoff}{A value between 0 and 1.  Calculations are
performed for values of LRO out to this percentile.  Optional.
The default value is 0.99.}

\item{Fdist}{The clutch size distribution.  The recognized options
are "Poisson" and "Bernoulli".  Optional.  The default value is
"Poisson".}
}
\value{
Returns a list containing the following:
\itemize{
\item probLifespanCondR: A matrix whose [i,j]th entry is the
probability that an individual with LRO = i-1 will have a
lifespan of j years (i.e. age j-1).
\item distKidsAtDeath: A vector whose jth entry is the probability of
having an LRO of j-1.
\item sdLifespanCondR: A vector whose jth entry is the standard
deviation of lifespan conditional on having an LRO of j-1.
\item CVLifespanCondR: A vector whose jth entry is the coefficient of
variation of lifespan conditional on having an LRO of j-1.
\item maxKidsIndex: calculations were performed out to values of LRO
equal to maxKidsIndex-1.
\item normalSurvProb: A vector whose jth entry is the unconditional
probability of surviving j years (i.e. until age j-1).
}
}
\description{
Calculates the distribution of lifespan conditional on LRO in the
presence of environmental variation.
}
\details{
Assumes a pre-breeding census (reproduction happens before
survival and growth).  The details of this calculation can be
found in Robin E. Snyder and Stephen P. Ellner.
2024. "To prosper, live long: Understanding the sources of
reproductive skew and extreme reproductive success in structured
populations."  The American Naturalist 204(2),
https://doi.org/10.1086/730557 and its online supplement.
}
\examples{
P1 = matrix (c(0, 0.3, 0, 0, 0, 0.5, 0, 0, 0.5), 3, 3)
P2 = matrix (c(0, 0.2, 0, 0, 0, 0.3, 0, 0, 0.4), 3, 3)
F1 = matrix (0, 3, 3); F1[1,] = 0:2
F2 = matrix (0, 3, 3); F1[1,] = 0.8*(0:2)
Plist = list (P1, P2)
Flist = list (F1, F2)
Q = matrix (1/2, 2, 2)
c0 = c(1,0,0)
maxClutchSize = 20
maxAge=20
out = distLifespanCondR2 (Plist, Flist, Q, c0, maxClutchSize, maxAge)
}
