% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distTraitCondR.R
\name{distLifespanCondR2NoEnv}
\alias{distLifespanCondR2NoEnv}
\title{Distribution of lifespan conditional on LRO}
\usage{
distLifespanCondR2NoEnv(
  Pmat,
  Fmat,
  c0,
  maxClutchSize,
  maxLRO,
  maxAge,
  percentileCutoff = 0.99,
  Fdist = "Poisson"
)
}
\arguments{
\item{Pmat}{The survival/growth transition matrix.
Pmat[i,j] is the probability of transitioning from
state j to state i.}

\item{Fmat}{The fecundity matrix.  Fmat[i,j]
is the expected number of state i offspring from a state j
parent.}

\item{c0}{A vector specifying the offspring state distribution:
c0[j] is the probability that an individual is born in state j}

\item{maxClutchSize}{The maximum clutch size to consider}

\item{maxLRO}{The maximum LRO to consider}

\item{maxAge}{The maximum attainable age}

\item{percentileCutoff}{A value between 0 and 1.  Calculations are
performed for values of LRO out to this percentile.  Optional.
The default value is 0.99.}

\item{Fdist}{The clutch size distribution.  The recognized options
are "Poisson" and "Bernoulli".  Optional.  The default value is
"Poisson".}
}
\value{
Returns a list containing the following:
\itemize{
\item probLifespanCondR: A matrix whose [i,j]th entry is the
probability that an individual with LRO = i-1 will have a
lifespan of j years (i.e. age j-1).
\item distKidsAtDeath: A vector whose jth entry is the probability of
having an LRO of j-1.
\item sdLifespanCondR: A vector whose jth entry is the standard
deviation of lifespan conditional on having an LRO of j-1.
\item CVLifespanCondR: A vector whose jth entry is the coefficient of
variation of lifespan conditional on having an LRO of j-1.
\item maxKidsIndex: calculations were performed out to values of LRO
equal to maxKidsIndex-1.
\item normalSurvProb: A vector whose jth entry is the unconditional
probability of surviving j years (i.e. until age j-1).
}
}
\description{
Calculates the distribution of lifespan conditional on LRO in the
absence of environmental variation.
}
\details{
Assumes a pre-breeding census (reproduction happens before
survival and growth).  The details of this calculation can be
found in Robin E. Snyder and Stephen P. Ellner.
2024. "To prosper, live long: Understanding the sources of
reproductive skew and extreme reproductive success in structured
populations."  The American Naturalist 204(2),
https://doi.org/10.1086/730557 and its online supplement.  This
is a wrapper function for distLifespanCondR2.
}
\examples{
Pmat = matrix (c(0, 0.3, 0, 0, 0, 0.5, 0, 0, 0.5), 3, 3)
Fmat = matrix (0, 3, 3); Fmat[1,] = 0:2
c0 = c(1,0,0)
maxClutchSize = 12
maxLRO = 30
maxAge=20
out = distLifespanCondR2NoEnv (Pmat, Fmat, c0, maxClutchSize,
  maxLRO, maxAge)
}
