% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_time_fns.R
\name{firstBreed}
\alias{firstBreed}
\title{Probability of breeding and age at first breeding}
\usage{
firstBreed(M0, pb)
}
\arguments{
\item{M0}{An \eqn{s \times s} matrix of transition rates among the \eqn{s}
non-breeding states in the population}

\item{pb}{A vector of length \eqn{s} containing the probability of breeding
(in the next time step) for each state.}
}
\value{
A list of:
\itemize{
\item \code{p_breed} A vector of length \eqn{s} containing the probability of breeding
before death, given the starting state.
\item \code{mean_age} A vector of length \eqn{s} containing the mean age of first
reproduction, given the starting state.
\item \code{var_age} A vector of length \eqn{s} containing the variance of age of first
reproduction, given the starting state.
}
}
\description{
Calculates the probability of breeding at least once, and the mean and
variance of age at first breeding conditional on breeding, as a function of
initial state. In order to use this function, the model specification must
include state-dependent breeding probability, not just state-dependent mean
fecundity. The probability of breeding at least once from now until death
must be positive for all states. If not, remove those states from the model
(can't start in any of them, going into them becomes death) and apply this
function to the remaining states.
}
\details{
In the \code{luckieR} package, we also provide the \code{probRepro()}
function, which can be used in scenarios where the distribution of
probability of breeding is not known. Instead, \code{probRepro} infers the
probability of breeding by assuming that the mean fecundity (provided in
\code{Fmat}) is the mean of a Poisson, Bernoulli, or fixed reproductive process.

Assumptions and coding follow Ellner et al. (2016) IPM monograph, chapter
3. The state transition matrix M is assumed to have the form M = p_b M_b +
(1-p_b) M_0 to allow for costs of reproduction, but M_b and M_0 can be
equal if there are no costs. Breeding does not necessarily imply producing
any new recruits, for example clutch size conditional on breeding could be
Poisson.
}
\examples{
Umat<- matrix(c(0.5, 0.1, 0.1, 0, 0.5, 0.3, 0, 0, 0.8), ncol=3)
pb<- c(0, 0.1, 0.4) # probability of breeding in the next time step, given current state
}
\seealso{
\code{\link[=probRepro]{probRepro()}} if you do not have values for \code{pb},
\code{\link[Rage:repro_maturity]{Rage::mature_age()}} which computes the mean age at first entering a state
with positive mean fecundity.
}
