% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megamatrixFunctions.R
\name{flatten}
\alias{flatten}
\title{Convert 4-d array to matrix}
\usage{
flatten(A4)
}
\arguments{
\item{A4}{The 4-d transition array}
}
\value{
The cross-classified transition matrix
}
\description{
Converts a 4-d transition array, such as the K array produced by
make_AxT, to a cross-classified transition matrix ("megamatrix"),
such as the A matrix produced by make_AxT
}
\examples{
M = matrix (c(0, 0.3, 0, 0, 0, 0.5, 0, 0, 0.5), 3, 3)
F = matrix (0, 3, 3); F[1,] = 0.1*(0:2)
B = mk_B (F, Fdist="Bernoulli")
mT = 6
out = make_AxT(B, M, mT)
K = out$K
A = flatten(out$K) ## should equal out$A
}
\seealso{
unfold, make_AxT
}
