% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_utilities.R
\name{fundamental_matrix}
\alias{fundamental_matrix}
\title{Calculate the fundamental matrix}
\usage{
fundamental_matrix(P)
}
\arguments{
\item{P}{An \eqn{n \times n} matrix of transition rates among the \eqn{n}
transient states in the matrix population model. If the absorbing state is
death, then this matrix is the \code{Umat} of transition rates among the living
stages}
}
\value{
An \eqn{n \times n} matrix where each \eqn{ij} entry is the expected
number of time steps spent in state \eqn{i} given that an individual starts
in state \eqn{j}
}
\description{
The fundamental matrix, \eqn{\mathbf{N}}, of a Markov Chain contains the
expected number of visits to each of the transient states before
transitioning into an absorbing state. It is given by:
\deqn{\mathbf{N}=(\mathbf{I}-\mathbf{P})^{-1},} where \eqn{\mathbf{P}} is the
matrix of transition probabilities among the transient states, and the power
\eqn{-1} indicates the matrix inverse.
}
\examples{
Umat<- matrix(c(0.5, 0.1, 0.1, 0, 0.5, 0.3, 0, 0, 0.8), ncol=3)
Nmat<- fundamental_matrix(Umat)
}
