% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megamatrixFunctions.R
\name{make_AxT}
\alias{make_AxT}
\title{Make the size x #kids transition matrix}
\usage{
make_AxT(B, M, mT)
}
\arguments{
\item{B}{The clutch size distribution matrix, where B[m,n] is the
probability that a size n parent produces m-1 offspring in a
single reproductive bout}

\item{M}{The state transition matrix.  States can be a single
quantity, such as size or life history stage, or can be
cross-classified, such as size x environment.  M[i,j] is the
probability of transitioning from state j to state i.}

\item{mT}{The dimension of the #kids part of the 4-d array.
I.e. the maximum #kids is mT-1.}
}
\value{
A list containing
\itemize{
\item A: The 2-dim transition matrix for states cross-classified by size
(or stage or...) and total number of offspring so far
\item K: The 4-dim transition array, with dimensions (next size, next
#kids, current size, current #kids).  K is modified so individuals
who get to the maximum values of #kids in the matrix stay there,
but continue to grow/shrink/die.
}
}
\description{
Makes the 2-dim iteration matrix A for a size x #kids model, where
#kids is the total number of offspring to date.
}
\details{
Called by distLifespanCondR2 and calcDistLRO.
}
\examples{
M = matrix (c(0, 0.3, 0, 0, 0, 0.5, 0, 0, 0.5), 3, 3)
Fmat = matrix (0, 3, 3); Fmat[1,] = 0:2
B = mk_B (Fmat)
mT = 20
out = make_AxT(B, M, mT)
}
