% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioningUtilities.R
\name{makeCondKernel}
\alias{makeCondKernel}
\title{calculates the probability of becoming successful before death and
the transition kernel conditional on becoming successful before
death.}
\usage{
makeCondKernel(M, transientStates)
}
\arguments{
\item{M}{The unconditional kernel (i.e. transition matrix)}

\item{transientStates}{A vector listing the transient
(non-absorbing) states}
}
\value{
A list containing
\itemize{
\item q2Extended: the probability of success
for all states.  The probability of success for the successful
states is 1 by definition.
\item MCond: The kernel conditional on becoming successful
}
}
\description{
Calculates the probability of hitting a set of absorbing states
before dying (i.e. before a stochastic trajectory generated by the
transition matrix ends).  If the set of absorbing states is defined
as "success," then this code calculates the probability of an
individual becoming successful before it dies.  Also calculates the
transition kernel conditional on becoming successful before death.
}
\examples{
M = matrix (0.1, 3, 3) ## States are small, medium, and large
## Transient states are small and medium.  "Success" is becoming
## large before you die.
transientStates = 1:2
out = makeCondKernel (M, transientStates)
}
