% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megamatrixFunctions.R
\name{makeM}
\alias{makeM}
\title{Make megamtrix}
\usage{
makeM(fastMatrixList, slowMatrix)
}
\arguments{
\item{fastMatrixList}{A list of transition matrices for the state
that cycles more rapidly in the cross-classified state.  There
should be one matrix for each state in the slow matrix.  e.g., if
the cross-classified state vector is (env. 1, stage 1; env. 1,
stage 2; ... env. 1, stage n; env. 2, stage 1, ...), then stage
is the fast state, and there should be a survival/growth matrix
for each environment or a fecundity matrix for each environment.}

\item{slowMatrix}{A transition matrix for the slow states in the
cross-classified state.  e.g., if the cross-classified state
vector is (env. 1, stage 1; env. 1, stage 2; ... env. 1, stage n;
env. 2, stage 1, ...), then environment is the slow state and
slowMatrix is the environment transition matrix.}
}
\value{
M, the transition matrix for the cross-classified state
}
\description{
Calculates the transition matrix for cross-classified states from
the transition matrices for each sub-classification:
i.e. calculates the megamatrix.
}
\examples{
P1 = matrix (c(0, 0.3, 0, 0, 0, 0.5, 0, 0, 0.5), 3, 3)
P2 = matrix (c(0, 0.2, 0, 0, 0, 0.3, 0, 0, 0.4), 3, 3)
F1 = matrix (0, 3, 3); F1[1,] = 0:2
F2 = matrix (0, 3, 3); F1[1,] = 0.8*(0:2)
Plist = list (P1, P2)
Flist = list (F1, F2)
Q = matrix (1/2, 2, 2)
M = makeM (Plist, Q)
bigF = makeM(Flist, Q)
}
