% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioningUtilities.R
\name{makeMCondLROThreshold}
\alias{makeMCondLROThreshold}
\title{Utility function for calculating a kernel conditional on having LRO
meeting or exceeding a threshold}
\usage{
makeMCondLROThreshold(
  M,
  Fmat,
  threshold,
  m0,
  maxLRO = 12,
  maxClutchSize = 12,
  Fdist = "Poisson"
)
}
\arguments{
\item{M}{the unconditional survival/growth transition matrix.
M[i,j] is the probability of surviving and transitioning from
state j to state i.}

\item{Fmat}{the fecundity matrix.  Fmat[i,j] is the expected number of
size i offspring from a size j parent.}

\item{threshold}{the number of offspring LRO should meet or exceed}

\item{m0}{the state distribution of offspring}

\item{maxLRO}{the maximum LRO is consider.  Optional, with a
default value of 12.}

\item{maxClutchSize}{the maximum number of offspring born in one
reproductive bout.  Optional, with a default value of 12.}

\item{Fdist}{the clutch size distribution.  Currently supported values are
"Poisson" and "Bernoulli."  Optional, with a default value of "Poisson".}
}
\value{
A list containing the following.  All matrices and vectors
are defined over an extended size distribution: see Details.
\itemize{
\item ACondSucceed: the size x #kids or size x env x #kids transition
matrix conditional on reaching the LRO threshold before dying
\item probSucceedCondZ: a vector whose jth entry is the probability
of reaching the threshold before dying, conditional on beginning
in size/stage j
\item probSucceed: the probability of succeeding assuming that an
individual's starting state is distributed according to m0
\item a0CondSucceed: the offspring state distribution conditional
on reaching the LRO threshold before dying.
}
}
\description{
Calculates the transition kernel conditional on producing at least
threshold offspring over the course of a life, as well as related
quantites.  Assumes a post-breeding census (survival and growth
happen before reproduction).
}
\details{
if there are n sizes/stages, states 1--n are those sizes
with LRO = 0 at the start of the time step (just after last
year's survival and growth (and environment) updates.  States
n+1--2n are those sizes with LRO = 0 just after this year's
fecundity update but before survival, growth (and environment)
updates.  States 2n+1--3n are those sizes with LRO = 1 at the
start of the time step, and so on through LRO = maxLRO.  To
advance the state by one time step, it is necessary to multiply
by ACondSucceed twice (once to update fecundity, then another
time to update survival, growth, and, if necessary, the
environment.
}
\examples{
M = matrix(0.1, 3, 3)
Fmat = matrix(0, 3, 3); Fmat[1,] = 0:2
threshold = 2
m0 = c(1, 0, 0)
out = makeMCondLROThreshold (M, Fmat, threshold, m0)
}
