% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioningUtilities.R
\name{makePCondBreedDef3}
\alias{makePCondBreedDef3}
\title{Utility function for calculating a kernel conditional on breeding.}
\usage{
makePCondBreedDef3(P, F, c0)
}
\arguments{
\item{P}{the survival/growth transition matrix}

\item{F}{the fecundity matrix.  The clutch size of a parent of size
j is Poisson-distributed with mean sum(F[,j]).}

\item{c0}{The size/stage distribution of offspring}
}
\value{
A list containing the following.  All matrices and vectors
are defined over an extended size distribution: if there are n
sizes/stages, states 1--n are those sizes with no offspring yet,
states n+1--2n are those sizes with the first offspring
produced in the current year, and states 2n+1--3n are those sizes
having produced at least one offspring in some past year.  See
Details.
\itemize{
\item PCondBreed: the transition matrix conditional on breeding before
dying
\item probEverBreedCondZ: a vector whose jth entry is the probability
of breeding before dying, conditional on beginning in size/stage j
\item PCondNeverBreedCondZ: the transition matrix conditional on not
breeding before dying
\item probEverBreedCondZ: a vector whose jth entry is the probability
of not breeding before dying, conditional on beginning in
size/stage j
\item bigc0CondBreed: the offspring size/stage distribution conditional
on the offspring breeding before dying.
\item bigc0CondNeverBreed: the offspring size/stage distribution conditional
on the offspring not breeding before dying.
\item pM: the unconditional probability of breeding
\item bigF: the fecundity matrix for the extended state space
}
}
\description{
Calculates the transition kernel conditional on producing at least
one offspring, as well as related quantites.  Assumes that the
clutch sizes are Poisson-distributed.
}
\details{
This function implements the calculations in Robin
E. Snyder, Stephen P. Ellner, Giles Hooker. 2021. "Time and Chance:
Using Age Partitioning to Understand How Luck Drives Variation in
Reproductive Success." The American Naturalist 197(4).
DOI: https://doi.org/10.1086/712874.  It assumes that a breeder is
an individual who has produced at least one offspring
("definition 3").
}
\examples{
P = matrix (c(0, 0.3, 0, 0, 0, 0.5, 0, 0, 0.5), 3, 3)
F = matrix (0, 3, 3); F[1,] = 0:2
c0 = c(1,0,0)
out = makePCondBreedDef3 (P, F, c0)
}
\seealso{
\link{makePDef3}, which this function calls
}
