% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioningUtilities.R
\name{makePDef3}
\alias{makePDef3}
\title{Utility function for calculating a kernel conditional on breeding.}
\usage{
makePDef3(P, F)
}
\arguments{
\item{P}{the survival/growth transition matrix.  P[i,j] is the
probability of surviving and transitioning from size j to size i.}

\item{F}{the fecundity matrix.  F[i,j] is the expected number of
size i offspring from a size j parent.}
}
\value{
bigP the extended state transition matrix (see Details)
}
\description{
This code is called by makePCondBreedDef3.  It assumes that clutch
sizes are Poisson-distributed.
}
\details{
This function calculates the extended state transition
kernel defined in eq. S34 of the supplementary information in Robin
E. Snyder, Stephen P. Ellner, Giles Hooker. 2021. "Time and Chance:
Using Age Partitioning to Understand How Luck Drives Variation in
Reproductive Success." The American Naturalist 197(4).
DOI: https://doi.org/10.1086/712874.  It assumes that a
breeder is defined as an individual who has produced at least one
offspring ("definition 3").
}
\examples{
P = matrix (c(0, 0.3, 0, 0, 0, 0.5, 0, 0, 0.5), 3, 3)
F = matrix (0, 3, 3); F[1,] = 0:2
out = makePDef3 (P, F)
}
\seealso{
\link{makePCondBreedDef3}, which calls this function
}
