% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_time_fns.R
\name{meanConditionalTimes}
\alias{meanConditionalTimes}
\title{Compute conditional time to arriving in a state}
\usage{
meanConditionalTimes(M)
}
\arguments{
\item{M}{An \eqn{n \times n} matrix of transition rates among the \eqn{n}
states in the matrix population model}
}
\value{
An \eqn{n \times n} matrix where the \eqn{(i,j)} element represents
the conditional mean time to reach i from j, with NA if the transition from
j to i is not possible.
}
\description{
Based on Equation 9 from Cochran and Ellner (1992), this function will
calculate the expected time for an individual to transition from a starting
state \eqn{j} to an arrival state \eqn{i}, conditional on arriving in state
\eqn{i}.
}
\examples{
Umat<- matrix(c(0.5, 0.1, 0.1, 0, 0.5, 0.3, 0, 0, 0.8), ncol=3)
meanCondTimes<- meanConditionalTimes(Umat)
}
