% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicLuckCalcs.R
\name{meanLifespan}
\alias{meanLifespan}
\title{Calculate mean lifespan}
\usage{
meanLifespan(Umat, mixdist = NULL)
}
\arguments{
\item{Umat}{An \eqn{n \times n} matrix of transition rates among the \eqn{n}
states in the matrix population model}

\item{mixdist}{Optional, a vector of length \eqn{n} that contains the
proportion of individuals starting in each state, to calculate a weighted
average lifespan over the possible starting states}
}
\value{
Value(s) of expected remaining lifespan in terms of number of model
timesteps. If \code{mixdist=NULL}, then \code{meanLifespan} returns a vector
containing the expected remaining lifespan for individuals starting in each
of the \eqn{n} states. If a mixing distribution is provided, then
\code{meanLifespan} returns a single value, calculated as a weighted average
over the offspring starting states.
}
\description{
Calculates the expected value of lifespan across individuals in a population
whose dynamics can be described a matrix population model. The survival and
transition rates among stages must be provided in the \code{Umat}. The default
behavior is to return the expected remaining lifespan for individuals
starting in all states of the population. This function also allows the user
to provide a "mixing distribution" of starting states. Note that in the
Markov Chain with Rewards framework, moments of lifespan are a special case
of moments of LRO, where individuals accrue fixed rewards of 1 for each
timestep that they survive.
}
\examples{
Umat<- matrix(c(0.5, 0.1, 0.1, 0, 0.5, 0.3, 0, 0, 0.8), ncol=3)
expectedVal<- meanLifespan(Umat)

# A standard choice for the mixing distribution is the distribution of
# offspring types in a cohort produced by the population at its stable
# distribution.
Fmat<- matrix(c(0.2, 0, 0, 1, 0.5, 0, 5, 3, 0), ncol=3)
Amat<- Umat+Fmat
mixdist<- calcDistOffspringCohort(Amat, Fmat)
expectedVal<- meanLifespan(Umat, mixdist=mixdist)
}
