% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicLuckCalcs.R
\name{meanLRO}
\alias{meanLRO}
\title{The mean of lifetime reproductive output (LRO)}
\usage{
meanLRO(Umat, Fmat, mixdist = NULL, offspring_weight = NULL)
}
\arguments{
\item{Umat}{An \eqn{n \times n} matrix of transition rates among the \eqn{n}
states in the matrix population model}

\item{Fmat}{An \eqn{n \times n} matrix of per capita rates of reproduction
rates by any of the \eqn{n} states into offspring (new individuals of any
of the \eqn{n} states)}

\item{mixdist}{Optional, a vector of length \eqn{n} that contains the
proportion of individuals starting in each state, to calculate a weighted
average LRO over the entire population}

\item{offspring_weight}{Optional, a vector of length \eqn{n} that contains
the relative value of offspring of different types. If included, the \code{Fmat}
is scaled column-wise by these weights, such that the rewards accrued
differ across offspring types. For more details and an example of usage,
see Hernandez et al. 2024. The natural history of luck: a synthesis study
of structured population models. Ecology Letters, 27, e14390. Available
from: https://doi.org/10.1111/ele.14390}
}
\value{
Value(s) of expected lifetime reproductive output. If
\code{mixdist=NULL}, then \code{meanLRO} returns a vector containing the expected
future LRO for individuals starting in each of the \exp{n} states. If a
mixing distribution is provided, then \code{meanLRO} returns a single value,
calculated as a weighted average of the offspring starting states.
}
\description{
Calculates the expected value of lifetime reproductive output, or lifetime
reproductive success, across individuals in a population whose dynamics can
be described a matrix population model. The survival and transition rates
among stages must be provided in the \code{Umat} and the mean per capita
reproductive output (per time step of the population model) must be provided
in the \code{Fmat}. The default behavior is to return the expected value of LRO
for individuals starting in all states of the population. This function also
allows the user to provide a "mixing distribution" of starting states, and/or
to provide offspring weights such that offspring of different types
contribute differently to reproductive success.
}
\examples{
Umat<- matrix(c(0.5, 0.1, 0.1, 0, 0.5, 0.3, 0, 0, 0.8), ncol=3)
Fmat<- matrix(c(0.2, 0, 0, 1, 0.5, 0, 5, 3, 0), ncol=3)
expectedVal<- meanLRO(Umat, Fmat)

# A standard choice for the mixing distribution is the distribution of
# offspring types in a cohort produced by the population at its stable
# distribution.
Amat<- Umat+Fmat
mixdist<- calcDistOffspringCohort(Amat, Fmat)
expectedVal<- meanLRO(Umat, Fmat, mixdist)

# In Hernandez et al. 2024, we used the probability of surviving to reproduce
# to rescale reproductive rewards when individuals can produce multiple
# offspring types.
off_wts<- probRepro(Umat, Fmat, repro_var='poisson')
expectedVal<- meanLRO(Umat, Fmat, offspring_weight=off_wts)
}
\seealso{
\code{\link[=calcDistOffspringCohort]{calcDistOffspringCohort()}}
}
