% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megamatrixFunctions.R
\name{mk_B}
\alias{mk_B}
\title{Make B matrix}
\usage{
mk_B(F, maxKids = 20, Fdist = "Poisson")
}
\arguments{
\item{F}{the fecundity matrix.  F[i,j] is the expected number of size i
offspring from a size j parent.}

\item{maxKids}{The maximum clutch size.  Optional, with a default value of
20.}

\item{Fdist}{the clutch size distribution.  Currently supported values are
"Poisson" and "Bernoulli."  Optional, with a default value of "Poisson".}
}
\value{
The clutch size distribution matrix B, where B[i,j] is the
probability that a size j parent produces i-1 offspring in a single
reproductive bout
}
\description{
Calculates B, the clutch size distribution matrix
}
\examples{
F = matrix (0, 3, 3); F[1,] = 0.1*(0:2)
out = mk_B (F, Fdist="Bernoulli")
}
