% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megamatrixFunctions.R
\name{mk_BPostBreeding}
\alias{mk_BPostBreeding}
\title{Make B matrix for a post-breeding census}
\usage{
mk_BPostBreeding(M, Fmat, maxKids = 20, Fdist = "Poisson")
}
\arguments{
\item{M}{The state transition matrix.  States can be a single
quantity, such as size or life history stage, or can be
cross-classified, such as size x environment.  M[i,j] is the
probability of transitioning from state j to state i.}

\item{Fmat}{the fecundity matrix.  Fmat[i,j] is the expected number of size i
offspring from a size j parent.}

\item{maxKids}{The maximum clutch size.  Optional, with a default value of
20.}

\item{Fdist}{the clutch size distribution.  Currently supported values are
"Poisson" and "Bernoulli."  Optional, with a default value of "Poisson".}
}
\value{
The clutch size distribution matrix B, where B[i,j] is the
probability that a size j parent produces i-1 offspring in a single
reproductive bout
}
\description{
Calculates B, the clutch size distribution matrix, for a
post-breeding census
}
\examples{
Fmat = matrix (c(0,0,0, 0.5,0,0, 0,0,0), 3, 3)
M = matrix (c(0,0.5,0, 0,0,0.5, 0,0,0), 3, 3)
maxClutchSize=10
out = mk_BPostBreeding (M, Fmat, maxClutchSize, Fdist="Bernoulli")
}
