% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megamatrixFunctions.R
\name{p_xT}
\alias{p_xT}
\title{Helper function used for calculating size x #kids transition
matrices}
\usage{
p_xT(l, i, j, B, Umat)
}
\arguments{
\item{l}{the value of total #kids in the next time step}

\item{i}{the current state}

\item{j}{the current total #kids}

\item{B}{The clutch size distribution matrix, where B[m,n] is the
probability that a size n parent produces m-1 offspring in a
single reproductive bout}

\item{Umat}{The state transition matrix.  States can be a single
quantity, such as size or life history stage, or can be
cross-classified, such as size x environment.  Umat[i,j] is the
probability of transitioning from state j to state i.}
}
\value{
A vector whose kth entry is the probability that a size i
individual with total number of offspring j will produce (l - j)
offspring in the current reproductive bout and will transition to
size k.  If l-j is < 0, the return value is a vector of zeros.
}
\description{
Function to take clutch size distribution matrix (B) and survival
and growth or survival and growth and environment transition matrix
(Umat), and compute the transition probabilities from size-class i and
j total kids, to all size classes and l total kids.
}
\details{
Called by makeAxT
}
\seealso{
makeAxT
}
