% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning.R
\name{partitionVarSkewnessEnvVar}
\alias{partitionVarSkewnessEnvVar}
\title{Partitions variance and skewness of LRO}
\usage{
partitionVarSkewnessEnvVar(
  Plist,
  Flist,
  Q,
  c0,
  maxAge = 100,
  survThreshold = 0.05,
  Fdist = "Poisson",
  esR1 = NULL,
  esR2 = NULL,
  esR3 = NULL,
  bsR1 = NULL,
  bsR2 = NULL,
  bsR3 = NULL
)
}
\arguments{
\item{Plist}{A list of survival/growth transition matrices.
Plist[[q]][i,j] is the probability of transitioning from state j to
state i in environment q.}

\item{Flist}{A list of fecundity matrices.  Flist[[q]][i,j] is the
expected number of state i offspring from a state j parent in environment q}

\item{Q}{The environment transition matrix.  Q[i,j] is the probability of
transitioning from environment j to environment i.}

\item{c0}{A vector specifying the offspring state distribution: c0[j] is
the probability that an individual is born in state j}

\item{maxAge}{The maximum age an individual can attain.  Optional. The
default value is 100.}

\item{survThreshold}{The threshold to use in determining lifespan (see return
values).  Optional.  Default value is 0.05.}

\item{Fdist}{The clutch size distribution.  The recognized options are
"Poisson" and "Bernoulli".  Optional.  The default value is "Poisson".}

\item{esR1}{The 1st order reward matrix for the extended state space (see
Details).  Optional.  The default assumes Poisson-distributed clutch sizes.}

\item{esR2}{The 2nd order reward matrix for the extended state space (see
Details).  Optional.  The default assumes Poisson-distributed clutch sizes.}

\item{esR3}{The 3rd order reward matrix for the extended state space (see
Details).  Optional.  The default assumes Poisson-distributed clutch sizes.}

\item{bsR1}{The 1st order reward matrix for the extended state space that
includes the pre-birth "ur-state" (see Details). Optional.  The default
assumes Poisson-distributed clutch sizes.}

\item{bsR2}{The 2nd order reward matrix for the extended state space that
includes the pre-birth "ur-state" (see Details). Optional.  The default
assumes Poisson-distributed clutch sizes.}

\item{bsR3}{The 3rd order reward matrix for the extended state space that
includes the pre-birth "ur-state" (see Details). Optional.  The default
assumes Poisson-distributed clutch sizes.}
}
\value{
A list containing the following: * birthStateVar: the contribution to
Var(LRO) from birth state luck * birthEnvVar: the contribution to Var(LRO)
from birth environment luck * survTrajecVar: a vector whose jth entry
contains the contribution to Var(LRO) from survival trajectory luck at age
j-1.  * growthTrajecVar: a vector whose jth entry contains the contribution
to Var(LRO) from growth trajectory luck at age j-1.
\itemize{
\item envTrajecVar: a vector whose jth entry contains the
contribution to Var(LRO) from environment trajectory luck at age j-1.  *
fecVar: a vector whose jth entry contains the contribution to Var(LRO) from
fecundity luck at age j-1.  * totVar: the total variance in LRO *
birthStateSkewness: the contribution to LRO skewness from birth state luck
\item birthEnvSkewness: the contribution to LRO skewness from birth environment
luck * survTrajecSkewness: a vector whose jth entry contains the
contribution to LRO skewness from survival trajectory luck at age j-1.  *
growthTrajecSkewness: a vector whose jth entry contains the contribution to
LRO skewness from growth trajectory luck at age j-1.  * envTrajecSkewness:
a vector whose jth entry contains the contribution to LRO skewness from
environment trajectory luck at age j-1.  * fecSkewness: a vector whose jth
entry contains the contribution to LRO skewness from fecundity luck at age
j-1.  * totSkewness: the total skewness in LRO * lifespan: the age by which
a proportion survThreshold of a cohort will be dead
}
}
\description{
Partitions variance and skewness of lifetime reproductive output (LRO) for
models with environmental variation but without trait variation.
}
\details{
A pre-breeding census is assumed: reproduction happens before
survival and growth.  The details of this calculation, including the
definitions of the extended states, can be found in Robin E. Snyder and
Stephen P. Ellner.  2024. "To prosper, live long: Understanding the sources
of reproductive skew and extreme reproductive success in structured
populations." The American Naturalist 204(2) and its online supplemnt.
}
\examples{
P1 = matrix (c(0, 0.3, 0, 0, 0, 0.5, 0, 0, 0.5), 3, 3)
P2 = matrix (c(0, 0.2, 0, 0, 0, 0.3, 0, 0, 0.4), 3, 3)
F1 = matrix (0, 3, 3); F1[1,] = 0:2
F2 = matrix (0, 3, 3); F1[1,] = 0.8*(0:2)
Plist = list (P1, P2)
Flist = list (F1, F2)
Q = matrix (1/2, 2, 2)
c0 = c(1,0,0)
out = partitionVarSkewnessEnvVar (Plist, Flist, Q, c0)
}
