% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning.R
\name{partitionVarSkewnessEnvVarAndTraits}
\alias{partitionVarSkewnessEnvVarAndTraits}
\title{Partitions variance and skewness of LRO}
\usage{
partitionVarSkewnessEnvVarAndTraits(
  PlistAllTraits,
  FlistAllTraits,
  Q,
  c0,
  traitDist,
  maxAge = 100,
  survThreshold = 0.05,
  Fdist = "Poisson",
  postBreedingCensus = FALSE
)
}
\arguments{
\item{PlistAllTraits}{A list of lists of survival/growth transition matrices.
Plist[[x]][[q]][i,j] is the probability of transitioning from
state j to state i in environment q when an individual has trait x.}

\item{FlistAllTraits}{A list of lists of fecundity matrices.
Flist[[x]][[q]][i,j] is the expected number of state i offspring
from a state j, trait x parent in environment q.}

\item{Q}{The environment transition matrix.  Q[i,j] is the probability of
transitioning from environment j to environment i.}

\item{c0}{A vector specifying the offspring state distribution: c0[j] is
the probability that an individual is born in state j}

\item{traitDist}{A vector whose jth entry is the probability of having trait
j}

\item{maxAge}{The maximum age an individual can attain.  Optional. The
default value is 100.}

\item{survThreshold}{The threshold to use in determining lifespan (see return
values).  Optional.  Default value is 0.05.}

\item{Fdist}{The clutch size distribution.  The recognized options are
"Poisson" and "Bernoulli".  Optional.  The default value is "Poisson".}

\item{postBreedingCensus}{Set this to TRUE to assume a post-breeding census
(i.e. reproduction happens after survival and growth).  Optional, with
default value FALSE.}
}
\value{
A list containing the following:
\itemize{
\item varFromTraits = the
contribution to Var(LRO) from trait variation
\item skewnessFromTraits = the contribution to LRO skewness from trait
variation * totVar: the total variance in LRO
\item totSkewness: the total skewness in LRO
\item birthEnvVar: the contribution to Var(LRO)
from birth environment luck
\item birthEnvSkewness: the contribution to LRO skewness from birth
environment luck * birthStateVar: the contribution to Var(LRO) from birth
state luck
\item birthStateSkewness: the contribution to LRO skewness from birth
state luck
\item survTrajecVar: a vector whose jth entry contains the
contribution to Var(LRO) from survival trajectory luck at age j-1.
\item survTrajecSkewness: a vector whose jth entry contains the
contribution to LRO skewness from survival trajectory luck at age j-1.
\item growthTrajecVar: a vector whose jth entry contains the
contribution to Var(LRO) from growth trajectory luck at age j-1.
\item growthTrajecSkewness: a vector whose jth entry contains the
contribution to LRO skewness from growth trajectory luck at age j-1.
\item envTrajecVar: a vector whose jth entry contains the
contribution to Var(LRO) from environment trajectory luck at age j-1.
\item envTrajecSkewness: a vector whose jth entry contains the
contribution to LRO skewness from environment trajectory luck at age j-1.
\item fecVar: a vector whose jth entry contains the
contribution to Var(LRO) from fecundity luck at age j-1.
\item fecSkewness: a vector whose jth entry contains the contribution
to LRO skewness from fecundity luck at age j-1.
\item totVarCondX: a vector whose jth entry is Var(LRO | trait = j)
\item totSkewnessCondX: a vector whose jth entry is the LRO skewness
conditional on trait = j
\item birthEnvVarCondX: a vector whose jth entry is the contribution of
birth environment luck to Var(LRO | trait = j)
\item birthEnvSkewnessCondX: a vector whose jth entry is
the contribution of birth environment luck to LRO skewness conditional on
trait = j
\item birthStateVarCondX: a vector whose jth entry is the contribution
of birth state luck to Var(LRO | trait = j)
\item birthStateSkewnessCondX: a vector whose jth entry is the
contribution of birth state luck to LRO skewness conditional on trait = j
\item survTrajecVarCondX: a matrix whose i,jth entry contains the
contribution to Var(LRO | trait = i) from survival trajectory luck at age
j-1.
\item survTrajecSkewnessCondX: a matrix whose i,jth entry contains the
contribution to LRO skewness conditional on trait = i from survival
trajectory luck at age j-1.
\item growthTrajecVarCondX: a matrix whose i,jth entry contains
the contribution to Var(LRO | trait = i) from growth trajectory luck at age
j-1.
\item growthTrajecSkewnessCondX: a matrix whose i,jth entry contains
the contribution to LRO skewness conditional on trait = i from growth
trajectory luck at age j-1.
\item envTrajecVarCondX: a matrix whose i,jth entry contains the
contribution to Var(LRO | trait = i) from environment trajectory luck at
age j-1.
\item envTrajecSkewnessCondX: a matrix whose i,jth entry contains the
contribution to LRO skewness conditional on trait = i from environment
trajectory luck at age j-1.
\item fecVarCondX: a matrix whose i,jth entry contains the contribution
to Var(LRO | trait = i) from fecundity luck at age j-1.
\item fecSkewnessCondX: a matrix whose i,jth entry contains the
contribution to LRO skewness conditional on trait = i from fecundity luck
at age j-1.
\item lifespanCondX: a vector whose jth entry is the age by which
survThreshold proportion of a cohort with trait j would be dead
}
}
\description{
Partitions variance and skewness of lifetime reproductive output (LRO) for
models with environmental variation and trait variation.
}
\details{
A pre-breeding census is assumed: reproduction happens before
survival and growth.  The details of this calculation, including the
definitions of the extended states, can be found in Robin E. Snyder and
Stephen P. Ellner.  2024. "To prosper, live long: Understanding the sources
of reproductive skew and extreme reproductive success in structured
populations." The American Naturalist 204(2) and its online supplement.
}
\examples{
PlistAllTraits = FlistAllTraits = list ()
P1 = matrix (c(0, 0.3, 0, 0, 0, 0.5, 0, 0, 0.5), 3, 3)
P2 = 0.9*P1
PlistAllTraits[[1]] = list (P1, P2)
P1 = matrix (c(0, 0.5, 0, 0, 0, 0.2, 0, 0, 0.7), 3, 3)
P2 = 0.9*P1
PlistAllTraits[[2]] = list (P1, P2)
F1 = matrix (0, 3, 3); F1[1,] = 0:2
F2 = 0.8*F1
FlistAllTraits[[1]] = list (F1, F2)
F1 = matrix (0, 3, 3); F1[1,] = 0.9*(0:2)
F2 = 1.1*F1
FlistAllTraits[[2]] = list (F1, F2)
Q = matrix (1/2, 2, 2)
c0 = c(1,0,0)
traitDist = rep(0.5, 2)
out = partitionVarSkewnessEnvVarAndTraits (PlistAllTraits,
      FlistAllTraits, Q, c0, traitDist)
}
