% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning.R
\name{partitionVarSkewnessNoEnvVarPostBreeding}
\alias{partitionVarSkewnessNoEnvVarPostBreeding}
\title{Partitions variance and skewness of LRO for a post-breeding census}
\usage{
partitionVarSkewnessNoEnvVarPostBreeding(
  Pmat,
  Fmat,
  c0,
  maxAge = 100,
  survThreshold = 0.05,
  Fdist = "Poisson",
  esR1 = NULL,
  esR2 = NULL,
  esR3 = NULL,
  bsR1 = NULL,
  bsR2 = NULL,
  bsR3 = NULL
)
}
\arguments{
\item{Pmat}{The survival/growth transition matrix: Pmat[i,j] is the probability
of transitioning from state j to state i.}

\item{Fmat}{The fecundity matrix: Fmat[i,j] is the expected number of state i
offspring from a state j parent}

\item{c0}{A vector specifying the offspring state distribution: c0[j] is
the probability that an individual is born in state j}

\item{maxAge}{The maximum age an individual can attain.  Optional. The
default value is 100.}

\item{survThreshold}{The threshold to use in determining lifespan (see return
values).  Optional.  Default value is 0.05.}

\item{Fdist}{The clutch size distribution.  The recognized options are
"Poisson" and "Bernoulli".  Optional.  The default value is "Poisson".}

\item{esR1}{The 1st order reward matrix for the extended state space (see
Details).  Optional.  The default assumes Poisson-distributed clutch sizes.}

\item{esR2}{The 2nd order reward matrix for the extended state space (see
Details).  Optional.  The default assumes Poisson-distributed clutch sizes.}

\item{esR3}{The 3rd order reward matrix for the extended state space (see
Details).  Optional.  The default assumes Poisson-distributed clutch sizes.}

\item{bsR1}{The 1st order reward matrix for the extended state space that
includes the pre-birth "ur-state" (see Details). Optional.  The default
assumes Poisson-distributed clutch sizes.}

\item{bsR2}{The 2nd order reward matrix for the extended state space that
includes the pre-birth "ur-state" (see Details). Optional.  The default
assumes Poisson-distributed clutch sizes.}

\item{bsR3}{The 3rd order reward matrix for the extended state space that
includes the pre-birth "ur-state" (see Details). Optional.  The default
assumes Poisson-distributed clutch sizes.}
}
\value{
A list containing the following: * birthStateVar: the contribution to
Var(LRO) from birth state luck * survTrajecVar: a vector whose jth entry
contains the contribution to Var(LRO) from survival trajectory luck at age
j-1.  * growthTrajecVar: a vector whose jth entry contains the contribution
to Var(LRO) from growth trajectory luck at age j-1.  * fecVar: a vector
whose jth entry contains the contribution to Var(LRO) from fecundity luck
at age j-1.  * totVar: the total variance in LRO * birthStateSkewness: the
contribution to LRO skewness from birth state luck * survTrajecSkewness: a
vector whose jth entry contains the contribution to LRO skewness from
survival trajectory luck at age j-1.  * growthTrajecSkewness: a vector
whose jth entry contains the contribution to LRO skewness from growth
trajectory luck at age j-1.  * fecSkewness: a vector whose jth entry
contains the contribution to LRO skewness from fecundity luck at age j-1.
\itemize{
\item totSkewness: the total skewness in LRO * lifespan: the age by which a
proportion survThreshold of a cohort will be dead
}
}
\description{
Partitions variance and skewness of lifetime reproductive output (LRO) for
models without trait or environmental variation, assuming a post-breeding
census.
}
\details{
A post-breeding census is assumed: reproduction happens after
survival and growth.  The details of this calculation, including the
definitions of the extended states, can be found in Robin E. Snyder and
Stephen P. Ellner.  2024. "To prosper, live long: Understanding the sources
of reproductive skew and extreme reproductive success in structured
populations." The American Naturalist 204(2) and its online supplement.
}
\examples{
Pmat = matrix (c(0, 0.3, 0, 0, 0, 0.5, 0, 0, 0.5), 3, 3)
Fmat = matrix (0, 3, 3); Fmat[1,] = 0:2
c0 = c(1,0,0)
out = partitionVarSkewnessNoEnvVarPostBreeding (Pmat, Fmat, c0)
}
