% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_time_fns.R
\name{popMeanVar}
\alias{popMeanVar}
\title{Population mean and variance of any attribute}
\usage{
popMeanVar(mean_by_type, var_by_type, mixdist)
}
\arguments{
\item{mean_by_type}{A vector of length \eqn{z} containing the mean of
attribute X for individuals of types 1 to \eqn{z}}

\item{var_by_type}{A vector of length \eqn{z} containing the variance of
attribute X for individuals of types 1 to \eqn{z}}

\item{mixdist}{A vector of length \eqn{z} indicating the frequency
distribution of individuals across the \eqn{z} types}
}
\value{
A list containing scalar values \code{pop_mean_X} and \code{pop_var_X} which
give the mean and variance, respectively, of attribute X across the
population
}
\description{
Computes the population mean and variance of some attribute X, given the
vectors of mean and variance conditional on individual 'type' Z. This
substitutes for all of the individual formulas in Table 2 of Cochran and
Ellner (1992). For example, if multiple offspring types are possible, this
can be used to calculate mean and variance over the possible starting states.
}
\examples{
Umat<- matrix(c(0.5, 0.1, 0.1, 0, 0.5, 0.3, 0, 0, 0.8), ncol=3)
expectedVal<- meanLifespan(Umat)
varVal<- varLifespan(Umat)
mixdist<- c(0.3, 0.7, 0)
popVals<- popMeanVar(expectedVal, varVal, mixdist)
}
\seealso{
\code{\link[=meanLifespan]{meanLifespan()}}, \code{\link[=meanLRO]{meanLRO()}}, \code{\link[=varLifespan]{varLifespan()}}, and \code{\link[=varLRO]{varLRO()}},
which should give the same results when the user provides a mixing
distribution.
}
