% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_time_fns.R
\name{popMu3}
\alias{popMu3}
\title{Population 3rd central moment of any attribute}
\usage{
popMu3(mean_by_type, var_by_type, mu3_by_type, mixdist)
}
\arguments{
\item{mean_by_type}{A vector of length \eqn{z} containing the mean of
attribute X for individuals of types 1 to \eqn{z}}

\item{var_by_type}{A vector of length \eqn{z} containing the variance of
attribute X for individuals of types 1 to \eqn{z}}

\item{mu3_by_type}{A vector of length \eqn{z} containing the third central
moment of attribute X for individuals of types 1 to \eqn{z}}

\item{mixdist}{A vector of length \eqn{z} indicating the frequency
distribution of individuals across the \eqn{z} types}
}
\value{
A scalar value for the 3rd central moment of attribute X across the
population
}
\description{
Function to compute the population third central moment of some attribute X,
given the vectors of mean, variance and mu3 conditional individual 'type' Z.
This is done using the order-3 case of the Law of Total Cumulance
}
\examples{
Umat<- matrix(c(0.5, 0.1, 0.1, 0, 0.5, 0.3, 0, 0, 0.8), ncol=3)
mean_by_type<- meanLifespan(Umat)
var_by_type<- varLifespan(Umat)
skew_by_type<- skewLifespan(Umat)
mu3_by_type = skew_by_type*((var_by_type)^(3/2))
mixdist<- c(0.3, 0.7, 0)
popMu3_Lifespan<- popMu3(mean_by_type, var_by_type, mu3_by_type, mixdist)
}
\seealso{
\code{\link[=meanLifespan]{meanLifespan()}}, \code{\link[=meanLRO]{meanLRO()}}, \code{\link[=varLifespan]{varLifespan()}}, and \code{\link[=varLRO]{varLRO()}},
which should give the same results when the user provides a mixing
distribution.
}
