% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_time_fns.R
\name{popSkew}
\alias{popSkew}
\title{Population skewness of any attribute X}
\usage{
popSkew(mean_by_type, var_by_type, skew_by_type, mixdist)
}
\arguments{
\item{mean_by_type}{A vector of length \eqn{z} containing the mean of
attribute X for individuals of types 1 to \eqn{z}}

\item{var_by_type}{A vector of length \eqn{z} containing the variance of
attribute X for individuals of types 1 to \eqn{z}}

\item{skew_by_type}{A vector of length \eqn{z} containing the skewness of
attribute X for individuals of types 1 to \eqn{z}}

\item{mixdist}{A vector of length \eqn{z} indicating the frequency
distribution of individuals across the \eqn{z} types}
}
\value{
A scalar value for the skewness of attribute X across the population
}
\description{
Function to compute the population skewness of some attribute X, given the
vectors of mean, variance and skewness conditional on individual 'type' Z.
Done by calling popMu3 and popMeanVar.
}
\examples{
Umat<- matrix(c(0.5, 0.1, 0.1, 0, 0.5, 0.3, 0, 0, 0.8), ncol=3)
mean_by_type<- meanLifespan(Umat)
var_by_type<- varLifespan(Umat)
skew_by_type<- skewLifespan(Umat)
mixdist<- c(0.3, 0.7, 0)
popSkew_Lifespan<- popSkew(mean_by_type, var_by_type, skew_by_type, mixdist)
}
\seealso{
\code{\link[=skewLifespan]{skewLifespan()}} and \code{\link[=skewLRO]{skewLRO()}}, which should give the same results
when the user provides a mixing distribution.

\code{\link[=popMu3]{popMu3()}} and \code{\link[=popMeanVar]{popMeanVar()}}
}
