% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicLuckCalcs.R
\name{probRepro}
\alias{probRepro}
\title{Probability of reproducing at least once, based on starting state}
\usage{
probRepro(Umat, Fmat, repro_var = "poisson")
}
\arguments{
\item{Umat}{An \eqn{n \times n} matrix of transition rates among the \eqn{n}
states in the matrix population model}

\item{Fmat}{An \eqn{n \times n} matrix of per capita rates of reproduction
rates by any of the \eqn{n} states into offspring (new individuals of any
of the \eqn{n} states)}

\item{repro_var}{The form of variance in reproductive output, with possible
values of \code{"poisson"}, \code{"bernoulli"} or \code{"fixed"}. This is required to
calculate the probability of breeding this year from the values of mean per
capita reproductive output.}
}
\value{
A vector of length \eqn{n} containing the probability of reproducing
at least once, for individuals starting in each of the \eqn{n} states
}
\description{
Calculates the probability of reproducing at least once, for individuals
starting in all possible starting states. In Hernandez et al. 2024, we used
the probability of surviving to reproduce to rescale reproductive rewards
when individuals can produce multiple offspring types.
}
\examples{
Umat<- matrix(c(0.5, 0.1, 0.1, 0, 0.5, 0.3, 0, 0, 0.8), ncol=3)
Fmat<- matrix(c(0.2, 0, 0, 1, 0.5, 0, 5, 3, 0), ncol=3)
probVals<- probRepro(Umat, Fmat, repro_var='poisson')
}
