% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distTraitCondR.R
\name{probTraitCondLRO}
\alias{probTraitCondLRO}
\title{Distribution of a trait conditional on LRO}
\usage{
probTraitCondLRO(
  PlistAllTraits,
  FlistAllTraits,
  Q,
  c0,
  maxClutchSize,
  maxLRO,
  traitDist,
  Fdist = "Poisson"
)
}
\arguments{
\item{PlistAllTraits}{A list of lists of survival/growth transition
matrices. Plist[[x]][[q]][i,j] is the probability of
transitioning from state j to state i in environment q when an
individual has trait x.}

\item{FlistAllTraits}{A list of lists of fecundity  matrices.
Flist[[x]][[q]][i,j] is the expected number of state i
offspring from a state j, trait x parent in environment q.}

\item{Q}{The environment transition matrix.  Q[i,j] is the
probability of transitioning from environment j to environment
i.}

\item{c0}{A vector specifying the offspring state distribution:
c0[j] is the probability that an individual is born in state j}

\item{maxClutchSize}{The maximum clutch size to consider}

\item{maxLRO}{The maximum LRO to consider}

\item{traitDist}{A vector whose jth entry is the unconditional
probability that the trait takes on the jth value}

\item{Fdist}{The clutch size distribution.  The recognized options
are "Poisson" and "Bernoulli".  Optional.  The default value is
"Poisson".}
}
\value{
A list containing the following:
\itemize{
\item probXCondR: a matrix whose i,jth component is the probability
that the trait has the jth value given that LRO = i-1
\item probR: A vector whose jth entry is the unconditional probability
that LRO = j-1
\item probRCondX: A matrix whose i,jth entry is the probability that
LRO = i-1 given that the trait takes the jth value.
}
}
\description{
Calculates Prob(X | R), where X is a trait value and R is lifetime
reproductive output (LRO).
}
\details{
The details of this calculation can be found in Robin
E. Snyder and Stephen P. Ellner.  2018.  "Pluck or Luck: Does Trait
Variation or Chance Drive Variation in Lifetime Reproductive
Success?"  The American Naturalist 191(4).  DOI:
10.1086/696125
}
\examples{
PlistAllTraits = FlistAllTraits = list ()
P1 = matrix (c(0, 0.3, 0, 0, 0, 0.5, 0, 0, 0.5), 3, 3)
P2 = 0.9*P1
PlistAllTraits[[1]] = list (P1, P2)
P1 = matrix (c(0, 0.5, 0, 0, 0, 0.2, 0, 0, 0.7), 3, 3)
P2 = 0.9*P1
PlistAllTraits[[2]] = list (P1, P2)
F1 = matrix (0, 3, 3); F1[1,] = 0:2
F2 = 0.8*F1
FlistAllTraits[[1]] = list (F1, F2)
F1 = matrix (0, 3, 3); F1[1,] = 0.9*(0:2)
F2 = 1.1*F1
FlistAllTraits[[2]] = list (F1, F2)
Q = matrix (1/2, 2, 2)
c0 = c(1,0,0)
traitDist = rep(0.5, 2)
out = probTraitCondLRO (PlistAllTraits, FlistAllTraits, Q,
      c0, maxClutchSize=10, maxLRO=15, traitDist)
}
