% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicLuckCalcs.R
\name{skewLRO}
\alias{skewLRO}
\title{Calculate skew in lifetime reproductive output}
\usage{
skewLRO(
  Umat,
  Fmat,
  repro_var = "poisson",
  mixdist = NULL,
  offspring_weight = NULL
)
}
\arguments{
\item{Umat}{An \eqn{n \times n} matrix of transition rates among the \eqn{n}
states in the matrix population model}

\item{Fmat}{An \eqn{n \times n} matrix of per capita rates of reproduction
rates by any of the \eqn{n} states into offspring (new individuals of any
of the \eqn{n} states)}

\item{repro_var}{The form of variance in per capita reproductive output, with
possible values of \code{"poisson"}, \code{"bernoulli"} or \code{"fixed"}.}

\item{mixdist}{Optional, a vector of length \eqn{n} that contains the
proportion of individuals starting in each state, to calculate a weighted
average LRO over the entire population}

\item{offspring_weight}{Optional, a vector of length \eqn{n} that contains
the relative value of offspring of different types. If included, the \code{Fmat}
is scaled column-wise by these weights, such that the rewards accrued
differ across offspring types. For more details and an example of usage,
see Hernandez et al. 2024. The natural history of luck: a synthesis study
of structured population models. Ecology Letters, 27, e14390. Available
from: https://doi.org/10.1111/ele.14390}
}
\value{
Value(s) of skewness in lifetime reproductive output. If
\code{mixdist=NULL}, then \code{skewLRO} returns a vector containing the skewness in
future LRO for individuals starting in each of the \eqn{n} states. If a
mixing distribution is provided, then \code{skewLRO} returns a single value of
skewness, calculated over the offspring starting states using the law of
total cumulance
}
\description{
Calculates the skew of lifetime reproductive output, or lifetime reproductive
success, across individuals in a population whose dynamics can be described a
matrix population model. The survival and transition rates among stages must
be provided in the \code{Umat} and the mean per capita reproductive output (per
time step of the population model) must be provided in the \code{Fmat}. The
default behavior is to assume Poisson-distributed annual reproduction among
individuals in a given stage class. The default is to return the skewness of
LRO for individuals starting in all states of the population. This function
also allows the user to provide a "mixing distribution" of starting states,
and/or to provide offspring weights such that offspring of different types
contribute differently to reproductive success.
}
\examples{
Umat<- matrix(c(0.5, 0.1, 0.1, 0, 0.5, 0.3, 0, 0, 0.8), ncol=3)
Fmat<- matrix(c(0.2, 0, 0, 1, 0.5, 0, 5, 3, 0), ncol=3)
skewness<- skewLRO(Umat, Fmat, repro_var='poisson')

# A standard choice for the mixing distribution is the distribution of
# offspring types in a cohort produced by the population at its stable
# distribution.
Amat<- Umat+Fmat
mixdist<- calcDistOffspringCohort(Amat, Fmat)
skew_overall<- skewLRO(Umat, Fmat, repro_var='poisson', mixdist=mixdist)

# In Hernandez et al. 2024, we used the probability of surviving to reproduce
# to rescale reproductive rewards when individuals can produce multiple
# offspring types.
off_wts<- probRepro(Umat, Fmat, repro_var='poisson')
skewness<- skewLRO(Umat, Fmat, repro_var='poisson', offspring_weight=off_wts)
}
