% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_utilities.R
\name{stable_dist}
\alias{stable_dist}
\title{Calculate the stable state distribution}
\usage{
stable_dist(Amat)
}
\arguments{
\item{Amat}{An \eqn{n \times n} matrix containing all of the transition and
reproductive rates that project the population from time \eqn{t} to
\eqn{t+1}. Generally given by \code{Umat+Fmat}.}
}
\value{
A vector of length \eqn{n} that contains the proportion of
individuals in each of the \eqn{n} states when the population is growing at
its asymptotic population growth
}
\description{
The stable state distribution is the proportional distribution of the
population across states (i.e., ages, stages, or sizes) when the population
is growing at its asymptotically-constant equilibrium growth rate
\eqn{\lambda}. The asymptotic population growth rate \eqn{\lambda} is given
by the eigenvalue with the largest real part. The stable state distribution
is the right eigenvector that corresponds to \eqn{\lambda}, normalized to sum
to 1.
}
\examples{
Umat<- matrix(c(0.5, 0.1, 0.1, 0, 0.5, 0.3, 0, 0, 0.8), ncol=3)
Fmat<- matrix(c(0.2, 0, 0, 1, 0.5, 0, 5, 3, 0), ncol=3)
Amat<- Umat+Fmat
ssd<- stable_dist(Amat)
}
