% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioningUtilities.R
\name{traitAve}
\alias{traitAve}
\title{Calculates an average over a user-provided distribution}
\usage{
traitAve(x, traitDist)
}
\arguments{
\item{x}{A vector}

\item{traitDist}{A vector containing the probability of each
value in x}
}
\value{
The average of x over traitDist
}
\description{
Calculates an average over a user-provided distribution
}
\examples{
x = 1:3
traitDist = rep (1/3, 3)
traitAve (x, traitDist)
}
