% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioningUtilities.R
\name{traitMu3}
\alias{traitMu3}
\title{Calculates the third central moment over a user-provided distribution}
\usage{
traitMu3(x, traitDist)
}
\arguments{
\item{x}{A vector}

\item{traitDist}{A vector containing the probability of each
value in x}
}
\value{
The third central moment of x over traitDist
}
\description{
Calculates the third central moment over a user-provided distribution
}
\examples{
x = 1:3
traitDist = rep (1/3, 3)
traitMu3 (x, traitDist)
}
