% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioningUtilities.R
\name{traitVar}
\alias{traitVar}
\title{Calculates a variance over a user-provided distribution}
\usage{
traitVar(x, traitDist)
}
\arguments{
\item{x}{A vector}

\item{traitDist}{A vector containing the probability of each
value in x}
}
\value{
The variance of x over traitDist
}
\description{
Calculates a variance over a user-provided distribution
}
\examples{
x = 1:3
traitDist = rep (1/3, 3)
traitVar (x, traitDist)
}
