% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megamatrixFunctions.R
\name{unfold}
\alias{unfold}
\title{Convert a megamatrix to an array}
\usage{
unfold(A2, dim)
}
\arguments{
\item{A2}{The megamatrix to be transformed}

\item{dim}{A vector containing the dimensions of the array to be returned}
}
\value{
The transition array
}
\description{
Converts a cross-classified transition matrix ("megamatrix") to a
4-d transition array, such as the K array produced by makeAxT
}
\examples{
Umat = matrix (c(0, 0.3, 0, 0, 0, 0.5, 0, 0, 0.5), 3, 3)
Fmat = matrix (0, 3, 3); Fmat[1,] = 0.1*(0:2)
B = makeB (Fmat, Fdist="Bernoulli")
mT = 6
out = makeAxT (B, Umat, mT)
A = out$A
K = unfold (A, dim=c(3,6,3,6))  ## should equal out$K
}
\seealso{
flatten, makeAxT
}
