% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megamatrixFunctions.R
\name{unvec}
\alias{unvec}
\title{Convert a vector to a matrix}
\usage{
unvec(nvec, nrow = NULL, ncol = NULL)
}
\arguments{
\item{nvec}{The vector to be turned into a matrix}

\item{nrow}{The number of rows in the matrix.  Either nrow or ncol
must be specified, but not both.}

\item{ncol}{The number of columns in the matrix.  Either nrow or ncol
must be specified, but not both.}
}
\value{
The matrix form of the vector
}
\description{
Converts a vector to a matrix with user-defined dimensions.
}
\details{
If the vector has length N and the desired number of rows
or columns is M, then N must be evenly divisible by M.  The
matrix is filled by columns.
}
\examples{
foo = 1:6
unvec (foo, nrow=2)
unvec (foo, ncol=2)
}
