% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_time_fns.R
\name{varConditionalTimes}
\alias{varConditionalTimes}
\title{Variance in conditional time to reach state i from state j}
\usage{
varConditionalTimes(M)
}
\arguments{
\item{M}{An \eqn{n \times n} matrix of transition rates among the \eqn{n}
states in the matrix population model}
}
\value{
An \eqn{n \times n} matrix where the \eqn{(i,j)} element represents
the variance in remaining expected lifespan of individuals starting in
state \eqn{j}, conditional on them reaching state \eqn{i} before death,
with NA if the transition from j to i is not possible.
}
\description{
Calculates the variance in lifespan for individuals starting in state
\eqn{j}, conditional on them reaching state \eqn{i} before death. This
calculation is based on Equation 10 in Cochran and Ellner (1992).
}
\examples{
Umat<- matrix(c(0.5, 0.1, 0.1, 0, 0.5, 0.3, 0, 0, 0.8), ncol=3)
varCondTimes<- varConditionalTimes(Umat)
}
