% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megamatrixFunctions.R
\name{vec}
\alias{vec}
\title{Convert a matrix to a vector}
\usage{
vec(nmat)
}
\arguments{
\item{nmat}{The matrix to be converted}
}
\value{
The vectorized form of the matrix
}
\description{
Converts a matrix to a vector by stacking columns
}
\examples{
foo = matrix (1:4, 2, 2)
vec (foo)
}
