% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioningUtilities.R
\name{makeMCondLROThreshold}
\alias{makeMCondLROThreshold}
\title{Utility function for calculating a kernel conditional on having LRO
meeting or exceeding a threshold}
\usage{
makeMCondLROThreshold(
  M,
  Fmat,
  threshold,
  m0,
  maxLRO = 12,
  maxClutchSize = 12,
  Fdist = "Poisson"
)
}
\arguments{
\item{M}{the unconditional survival/growth transition matrix.
M[i,j] is the probability of surviving and transitioning from
state j to state i.}

\item{Fmat}{the fecundity matrix.  Fmat[i,j] is the expected number of
size i offspring from a size j parent.}

\item{threshold}{the number of offspring LRO should meet or exceed}

\item{m0}{the state distribution of offspring}

\item{maxLRO}{the maximum LRO is consider.  Optional, with a
default value of 12.}

\item{maxClutchSize}{the maximum number of offspring born in one
reproductive bout.  Optional, with a default value of 12.}

\item{Fdist}{the clutch size distribution.  Currently supported values are
"Poisson" and "Bernoulli."  Optional, with a default value of "Poisson".}
}
\value{
A list containing the following.  All matrices and vectors
are defined over an extended size distribution: if there are n
sizes/stages, states 1--n are those sizes with no offspring yet,
states n+1--2n are those sizes with the first offspring
produced in the current year, and states 2n+1--3n are those sizes
having produced at least one offspring in some past year.  See
Details.
\itemize{
\item ACondSucceed: the a size x #kids or size x env x #kids transition
matrix conditional on reaching the LRO threshold before dying
\item probSucceedCondZ: a vector whose jth entry is the probability
of reaching the threshold before dying, conditional on beginning
in size/stage j
\item m0CondSucceed: the offspring state distribution conditional
on reaching the offspring reaching the threshold before dying.
}
}
\description{
Calculates the transition kernel conditional on producing at least
threshold offspring over the course of a life, as well as related
quantites.
}
\details{
To do this calculation, we extend the state space so that
number of offspring is added to the state definition, and all
outputs are given for this extended state space.  For
example, if the original state space was (size 1, size 2),
the new state space is (size 1, 0 offspring; size 2, 0 offspring;
size 1, 1 offspring; size 2, 1 offspring, ...) up to #offspring =
maxLRO.
}
\examples{
M = matrix(0.1, 3, 3)
Fmat = matrix(0, 3, 3); Fmat[1,] = 0:2
threshold = 2
m0 = c(1, 0, 0)
out = makeMCondLROThreshold(M, Fmat, threshold, m0)
}
